# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Genesis 2.0",
    "description": "Endpoints Genesis 2.0",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://tudominio.com/api",
      "description": "Servidor de producción"
    }
  ],
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/spaces/reserve": {
      "post": {
        "tags": [
          "Space Reservations"
        ],
        "summary": "Reserve a space",
        "description": "Creates a new space reservation with the specified details including space ID, user ID, dates and reason",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "spaceId": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "startDate": {
                  "example": "any"
                },
                "endDate": {
                  "example": "any"
                },
                "reason": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spaces/history/{userId}": {
      "get": {
        "tags": [
          "Space Reservations"
        ],
        "summary": "Get user reservation history",
        "description": "Retrieves the reservation history for a specific user with pagination support",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spaces/spaces": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "summary": "Get all available spaces",
        "description": "Retrieves a list of all available spaces with their IDs and names",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spaces/all-reservations": {
      "get": {
        "tags": [
          "Space Reservations"
        ],
        "summary": "Get all reservations",
        "description": "Retrieves all space reservations with pagination support",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spaces/update-status/{reservaId}": {
      "put": {
        "tags": [
          "Space Reservations"
        ],
        "summary": "Update reservation status",
        "description": "Updates the status of a specific reservation",
        "parameters": [
          {
            "name": "reservaId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newStatus": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/spaces/update-reservation/{reservaId}": {
      "put": {
        "tags": [
          "Space Reservations"
        ],
        "summary": "Update reservation details",
        "description": "Updates the details of a specific reservation including dates, reason and space",
        "parameters": [
          {
            "name": "reservaId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "newStartDate": {
                  "example": "any"
                },
                "newEndDate": {
                  "example": "any"
                },
                "newReason": {
                  "example": "any"
                },
                "newSpaceId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/spaces/reservation/{reservaId}": {
      "get": {
        "tags": [
          "Space Reservations"
        ],
        "summary": "Get reservation details",
        "description": "Retrieves detailed information about a specific reservation",
        "parameters": [
          {
            "name": "reservaId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}